user: -changed wokeinterval to an int settings_notifs: -added a seekbar -updates the user's wokeinterval -changed layout to 20dp margin settings_home: -changed layout to 20dp margin -made imagebutton_notif extend all the way to edge of the screen -made selector for imagebutton press studysession: -made onTick() display minutes/secs left till next notif from beginning of study session

+++ app/src/main/java/com/pineapple/woke/resources/User.java
@@ -10,11 +10,11 @@ import java.util.ArrayList;
 public class User {
 
     private String name;
-    private double wokeInterval;
+    private int wokeInterval;
     private ArrayList<SavedSession> studySessions;
 
     public User(){ }
-    public User(String name, double wokeInterval){
+    public User(String name, int wokeInterval){
         this.name = name;
         this.wokeInterval = wokeInterval;
         studySessions = new ArrayList<>();
@@ -27,10 +27,10 @@ public class User {
         this.name = name;
     }
 
-    public double getWokeInterval() {
+    public int getWokeInterval() {
         return wokeInterval;
     }
-    public void setWokeInterval(double wokeInterval) {
+    public void setWokeInterval(int wokeInterval) {
         this.wokeInterval = wokeInterval;
     }
